+++ src/ClearDashboard.DAL/Models/ParallelVerse.cs
@@ -5,10 +5,10 @@ namespace ClearDashboard.DataAccessLayer.Models
 {
     public partial class ParallelVerse
     {
-        public long Id { get; set; }
-        public long SourceVerseId { get; set; }
-        public long TargetVerseId { get; set; }
-        public long? ParallelCorpusId { get; set; }
+        public int Id { get; set; }
+        public int SourceVerseId { get; set; }
+        public int TargetVerseId { get; set; }
+        public int? ParallelCorpusId { get; set; }
 
         public virtual ParallelCorpus ParallelCorpus { get; set; }
         public virtual Verse VerseVerse1 { get; set; }
